-pressing enter when adding a new platform also calls the function to create it
-resizing stuff
-"additional comments" so you know for what account it is
-logo scraper was changed to requests fully, and uses an api by clearbit to fetch logos instead of googling
-progress bar for when scraping logos after backup
-backups now accept user strings
-scraping now asynchronous
-changed the font for generated platform buttons
-added common logos, as to not scrape if someone simply wants ie discord
-You can now change the image of a platform
-You can now delete a platform
-Made an asynchronous version of the logo scraper, so when you read from a backup file it finds the logos for all the platforms
-everytime you generate a new password, it saves automatically into passwords.json, preventing you from losing data if the program crashes

If i missed anything its in the commit history of compare-the-differences

+++ password manager/python scraper/scraper.py
@@ -2,23 +2,38 @@ import selenium
 import sys
 from selenium import webdriver
 from selenium.webdriver.common.by import By
+from selenium.webdriver.chrome.options import Options
 import requests
 import time
+import os
 
 def scrapeLogo(inputString):
     print("herher")
-    webDriving = webdriver.Chrome()
-    searchString = "https://duckduckgo.com/?q=" + inputString + "+logo+png"
+
+    #Web driver options to make it invisible
+    webDriverOptions = Options()
+    #webDriverOptions.add_argument("--headless")
+    #webDriverOptions.add_argument("--disable-gpu")
+
+
+    webDriving = webdriver.Chrome(options=webDriverOptions)
+    searchString = "https://duckduckgo.com/?q=" + inputString + "+transparent+logo+png&ia=images"
     webDriving.get(searchString)
 
+    #Wait half a second to grab the images anchor tag
+    time.sleep(0.5)
+    imageTab = webDriving.find_element(By.LINK_TEXT, "Images")
+    imageTab.click()
+
+    #wait half a second for the images to load
+    time.sleep(0.5)
+
     imgLink = None
-    starttime = time.time_ns()
-    currentTime = starttime
-    timeout = 5 * (10 ** (-9))
+
     print("here")
-    while imgLink == None and currentTime - starttime < timeout:
+    while imgLink == None:
         imgLink = webDriving.find_element(By.XPATH, "//img[@loading = 'lazy']")
-        currentTime = time.time_ns()
+        time.sleep(0.001)
         
     if imgLink == None:
         print("nema slike :(((((((")
@@ -28,11 +43,40 @@ def scrapeLogo(inputString):
 
     imgWritePath = "img\\logos\\" + inputString + ".png"
     print(imgWritePath)
-    with open(imgWritePath, "wb") as file:
-        file.write(imageSrc.content)
+    if not os.path.exists(imgWritePath):
+        with open(imgWritePath, "wb") as file:
+            file.write(imageSrc.content)
+    else:
+        print("There is already an image here, no need to write one again")
 
-    time.sleep(3)
+    time.sleep(1)
 
     webDriving.quit()
 
-scrapeLogo(sys.argv[1])
+def scrapeLogoClearbit(inputString): #NEW AND IMPROVED SCRAPER
+    inputString = str.lower(inputString)
+    imgWritePath = "img\\logos\\" + inputString + ".png"
+
+    if os.path.exists(imgWritePath):
+        print("There is already an image here, no need to write one again.")
+        return imgWritePath
+
+    #Clearbit has a logo api that can be scraped 50k times a month for free per the docs, doubt any user would exceed that number 
+    searchString = "https://logo.clearbit.com/" + inputString + ".com"
+    
+    scraperRequest = requests.get(searchString, stream=True)
+
+    if scraperRequest.status_code == 200:
+        os.makedirs(os.path.dirname(imgWritePath), exist_ok=True)
+
+        with open(imgWritePath, "wb") as scrapedImg:
+            for byteChunks in scraperRequest.iter_content(1024):
+                if byteChunks:
+                    scrapedImg.write(byteChunks)
+
+        print("Logo saved at " + imgWritePath)
+    else:
+        print(f"HTTP ERROR CODE {scraperRequest.status_code}")
+
+
+scrapeLogoClearbit(sys.argv[1])
